Rename geojson module to geom to avoid confusion with geojson crate
